Only write a JS file if the target is a file.

+++ src/de/inetsoftware/jwebassembly/module/WasmTarget.java
@@ -138,7 +138,7 @@ public class WasmTarget implements Closeable {
      *             if any I/O error occur
      */
     public Writer getJavaScriptOutput() throws IOException {
-        if( javaScript == null && file != null ) {
+        if( javaScript == null && file != null && file.isFile() ) {
             javaScript = new OutputStreamWriter( new BufferedOutputStream( new FileOutputStream( file + ".js" ) ), StandardCharsets.UTF_8 );
         }
         return javaScript;
